Support UTCTiming element with the http-xsdate scheme
While both doing the inventory of what MPD attributes handle the RxPlayer
and working on low-latency contents, I noted that the RxPlayer did not
support `UTCTiming` element with the
`urn:mpeg:dash:utc:http-xsdate:2014`.

Yet it is very simple to implement - at least in JavaScript - as we
already handle the `http-iso` format and that both are valid ISO8601
dates. As we use the `Date` builtin to transform such strings to a JS
`Number`, we can thus use the exact same parsing logic for both.