Reduce coupling between fetchers and ABR by using callbacks in the former
This commit tries to reduce the strong implicit link between the
`fetchers` (in `src/core/fetchers`) and the ABR logic (in
`src/core/abr`).

Why
===

In the previous code, the `SegmentFetcher` (deep inside the fetchers
code) directly used an RxJS Subject to send various network lifecycle
events which was listened to by the `RepresentationEstimator` (deep
inside the ABR code). The `stream` (in `src/core/stream`) was the one
linking both by creating the Subject and passing it to both modules.

I wasn't happy with that link, even more considering that new cases (the
`VideoThumbnailLoader` and the future `ContentDownloader`) may want to
use the `fetchers` code to load and parse segments without needing any
adaptive logic from the ABR. In that case, they needed to provide a mute
Subject, which didn't make much sense.

What
====

That's the reason why I decided to switch to a more generic
callback-based solution instead. Instead of a Subject on which various
`fetchers` lifecycle events (such as: `requestBegin` or `requestEnd`)
will be sent, the caller can now provide callbacks to be notified on
those various events.

This way of doing it is more usual in JS-land and also goes toward our
long term goal of phasing out the hard-to-grasp RxJS to more usual
patterns, to improve the RxPlayer code readability and debuggability.
It also improve the code of the `AdaptationStream`, which had to perform
some tricks to perform the link between the abr and fetchers code.

I also profited from the modification here to continue our recent style
change of putting TypeScript interface definitions after the main export
to improve readability for newcomers. A huge part of the diff may be due
to that.

This is still not perfect though as data communicated through the
callbacks still respect an interface coming from the ABR. I chose this
for simplicity reasons.