update dependencies and code
This is the usual dependency update before a release (scheduled
to be the v3.26.2).

After updating `typescript-eslint`, new linter errors made their
apparitions, for cases where an `any` type is used as a function
argument (the rule in question is
[no-unsafe-argument](https://github.com/typescript-eslint/typescript-eslint/blob/main/packages/eslint-plugin/docs/rules/no-unsafe-argument.md)).

That's a rule we want (`any` is ___bad___ after all!) but RxJS,
our main dependency, does not respect it well.
Especially, some operators which transform an observable into
another without needing to look at the items of the source
Observable (e.g. `mapTo`, `mergeMapTo`, `switchMapTo`,
`ignoreElements`) have an `any` in their type definition for the type
of the source Observable's elements (where an `unknown` would actually
suffice and be type-safe).

Replacing on their side this `any` by `unknown` is planned for their
8.x.x release (which is not yet released). In the meantime, we had to
find a solution.

I chose to do two things:

  1. Replace all the "***To" (`mapTo`, `mergeMapTo` and so on)
     operators by their callback-based equivalent (`map`,
     `mergeMap` and so on).
     I did that mostly because RxJS devs [seems to plan
     deprecating the ***To operators anyway in profit of the
     callback-based ones](https://github.com/ReactiveX/rxjs/issues/6399).

  2. For the last remaining operator, `ignoreElements()`, I just
  disabled the rule with a long comment attached on top of each case.