Better estimate the duration of ISOBMFF segments with multiple moof boxes
This is a commit with a smallish improvement to better estimate the
duration of segments with multiple "moof" boxes (for example, low-latency
segments loaded as a whole instead of per-chunk).

Previously, only the first "trun" box of the first "traf" box of the
first "moof" box encountered in an ISOBMFF segment was parsed to obtain
the segment's duration.
In most cases, this works. However in some rare situations, media
segments might actually contain several subsegments, in which case we
would only calculate the first of those subsegments' duration.

Now, we always check if multiple moof boxes are there and, if it is the
case, we addition the duration calculated from all found trun boxes.

The previously miscalculted duration never provoked any sensible issue
because we have a safeguard preventing huge differences between the
duration anounced in the MPD and what has actually been parsed from the
segment.
In that case, we used the segment duration as anounced in the MPD, which
may be less precise that the segment's metadata (especially if e.g. this
is a `$Number$`-based MPD content.

I just tested it with success on some low-latency contents.