Fix native track elements' content not removed in Chrome 96
This commit works around an issue seen recently with Chrome 96 only
where native text tracks would not be visually removed after stopping a
content / reloading another.

First thought to be an RxPlayer issue, we noticed that the corresponding
`<track>` element was actually properly removed from the DOM and that
the issue only appeared in Chrome, even on Edge Chromium things worked
as intented.

After investigation, the issue seems to be linked to Chrome: when the
`<track>` element is removed from the DOM synchronously after (or in a
microtask after) emptying the src attribute of the linked media
element, the `<track>` element disappears from the DOM, but its content
stays on it.

Funnily enough, it seems (I forgot about it) that the same issue was
present in FireFox some years ago, judging by some firefox-specific code
in our `compat` code. Now FireFox doesn't seem to have the issue anymore
but Chrome does! `<track>` element handling seems to be a bug-prone area
of browsers!

To fix that issue, I chose to generalize what was before a
FireFox-specific issue:
When the `src` attribute of the media element is emptied, we first
remove all associated `<track>` elements from it.

This sadly means that an application might not be able to handle
`<track>` element by itself (for example, to manually add subtitles),
though I still went on with it because:
  1. This should be an extremely rare usecase, even more now that we
     have more powerful solutions such as the `TextTrackRenderer` tool.
  2. Updating data on the HTMLMediaElement while it is linked to the
     RxPlayer is already a risky behavior, where multiple things can
     behave not as intented.