scripts: `check` script does not run unit tests anymore
The `check` script, runnable through `npm run check`, previously ran:
  1. TypeScript to check types in the src directory
  2. eslint to lint the src directory
  3. All unit tests

The goal of that script was that it could be used before all commits /
push, so we can avoid pushing mistakes when in the review phase.

However, I suspect that the fact that it runs all unit tests, which can
be a long operation (43 seconds for me on the current master), reduce its
attractiveness enough that we often skip that step.

As unit tests are run by our CI (and often manually) anyway, I'm
removing it from the `npm run check` script which will now only type
check and lint, which are both the most useful and the faster operations
of the three.