DASH: always consider that the non-last Period is finished in SegmentTimeline contexts
This commit fixes an issue we recently had at Canal+ linked to ambiguous
DASH Period boundaries:

An infinite rebuffering issue could arise when playing a live
multi-Period DASH content where the audio and video tracks had different
ending timestamp (even in the order of some milliseconds) AND the
duration of the Period (either taken from the `Period@duration` attribute
or inferred from the `Period@start` of the next Period if one) was set to
be the maximum time between them (which very oddly is the DASH-IF
recommendation).

In that condition, the RxPlayer could be waiting indefinitely at the end
of the corresponding Period for what it thinks to be the last - but is
actually inexistent - audio or video segment (respectively depending on
if the audio track or the video track is shorter)  to be generated and
that even as new Periods are defined in the MPD.

We already were aware of this possibility but thought that for multiple
reasons it made more sense to always set the Period's duration to the
minimum between all tracks, as it leads to less ambiguity especially when
we're dealing with `<SegmentTemplate>` indexing without a
`<SegmentTimeline>`, in which case there would be no way to know which
segment is the last one for each track.

In a case with a `<SegmentTimeline>` though, this issue is theoretically
resolvable.
If we consider that as long the corresponding Period is not the last one
in the MPD, it means that its segments have been fully generated.

We were for a long time unsure of if it was the case, so we were very
reticent to do that development.
What changed things very recently however, is that we just encountered
the case - and the concerned people don't seem to want to remove that
ambiguity on the packaging side for various reasons - and that the
`dash.js` player seems to have taken the decision [to do exactly the
fix we're speaking of](https://github.com/Dash-Industry-Forum/dash.js/blob/01b27851514e6a3dcd9c23743f789c601de6fde7/src/dash/DashHandler.js#L225-L228)
(_trigger warning (for the link): veeery long lines!_).

I'm however still afraid of breaking things for (arguably extremely
rare) use cases where, e.g., the penultimate Period would still be in a
process of being generated (in which case, we will skip it as soon as we
reach its current end).

---

As this commit add yet another property in the context objects
communicated in the DASH MPD parser function (more specifically, in the
common part between the JS and WASM parser), most of the diff of that
commit is actually linked to DRYing the linked type definitions - with
the help of TypeScript builtin `Omit` type, so that they are only
documented once in the code.
And as it is the new way of doing things, I also moved type definitions
that I updated at the end of the corresponding files (so that what a new
dev sees first is the actual code, not types).