Avoid warning logs when "creating" initialization segments
A small oversight led to the RxPlayer potentially logging a warning each
time a segment is either created directly from the Manifest's
information - which as far as I remember - only happens in Smooth
contents or when no initialization segment exist (such as for
subtitles).

Aside from this unnecessary warning, there should be no other real
issue.

This warning was due to the fact that we're anouncing the end of those
initialization's "requests" (between quotes as there there is in fact no
request happening) two times in a row, breaking an assertion later in
the code that ended request should be pending in the first place.

It's annoying that I saw that just after finalizing the release, but
thankfully this is a very minor issue that will probably not bother
anyone, and should appear relatively rarely for DASH contents (maybe once
when switching to a track with subtitles? Even this is unsure).