add forgotten TypeScript types in the exposed segment and manifest loader APIs
An application developper recently let me know that the type definitions
for the RxPlayer's `transportOptions.segmentLoader` and
`transportOptions.manifestLoader` loadVideo option were missing some
attributes/arguments:
  - the `url` property that a manifest loader can communicate to the
    RxPlayer in case of a redirection is not defined (even if setting it
    worked).
  - the `progress` callback provided to the segment loader to notify
    about the request was not defined (even if using it worked)

I also noticed a third issue, the `fallback` callback allowing to
fallback on the RxPlayer's default loader implementation could be
undefined. I guess it could make sense in some cases, such as cases
where the RxPlayer is not able to do requests (for example: a
"local-manifest" content) but in that case we would not be aligned
with what the API documentation says, as it doesn't say anything about
that callback's optionality. And anyway, just throwing in those very
rare situations when we cannot fallback should not be that problematic.