only rely on pinkie's Promise when necessary
While doing some debugging on an unrelated matter, I noticed that
pinkie's Promise implementation was being used even on a recent browser
with a native Promise implementation.

This is due to mix-ups in the code, where we're including directly this
implementation (`import PPromise from "pinkie";`) instead of importing
the file selecting one or the other depending on browser support
(`import PPromise from "../../utils/promise"`).

This is not a big issue at all, and I would guess most places still
relied on the native implementation, but it's probably better for
performance reason to use the native one when available.

I was also wondering if it made sense to use pinkie at all. Originally,
it was just imported so the player can be used on IE11.
Yet it seems that RxJS, one of our other dependencies, does not
actually work out-of-the box on IE11 without a Promise polyfill already
declared by the application.

I didn't remove the pinkie dependency for now, but I would be very ready
to do so if we can prove that this is completely unnecessary.