Remove "sanitize-html" dependency
After receiving a security alert and assessing the real risk for the
project, I saw that we were using the "sanitize-html" dependency on some
scripts despite it being particularly overkill for the needed case.

What we wanted to actually do instead when we used that dependency was
just to escape either some Node's inner text and some attribute values.
If I understand correctly what we need to do is just to replace
characters by the corresponding html entity. We already depend on the
`html-entities` which does this for us, and isn't as large as
"sanitize-html" (the former also has 0 dependency) and doesn't do
unneeded things like processing CSS.

So here I removed the "sanitize-html" dependency, replacing it with
entity-replacing code.