DASH: properly update `startNumber` on SegmentTimeline GC
Fixes #1052.

We had an issue with our logic which automatically removes outdated
references to segments the RxPlayer found in DASH' SegmentTimeline.

After removing segments behind the current DVR window (a.k.a. timeshift
buffer depth), it wasn't updating the potentially-linked `startNumber`
attribute, which is useful in case a `$Number$` URL scheme is used
(thankfully this is rare with DASH SegmentTimeline, where `$Time$` is
usually preferred in that case).

Due to this a bug could be encountered when the beginning of that
`$Number$`-based SegmentTimeline was before the DVR window. In that
case, the wrong segment's number would be calculated and thus, the wrong
segment would be loaded.

This commit fixes the issue by making all functions that may remove
elements from a referenced `timeline` object returns a way to determine
how much segments have been removed (including implicit segments which
are e.g. implied by a SegmentTimeline's `r` attribute).