directfile: disable all non-enabled audio tracks
We recently observed an issue on Safari where enabling an audio track
in directfile mode without disabling the others led to the audio track
not being audible.

To our knowledge, Safari had no problem with this previously but by
checking the HTML5 living standard, we can see that at least now,
enabling an audio track does not mean to disable the others and could
even mean having multiple audio tracks at once.

This commit thus disable all other audiotracks when enabling a new one,
which should not be problematic in any case.

It has been tested on Safari with success.