text: fix subtitles skipping issue on overlapping cues
We discovered an issue where subtitles could be skipped in rare
situations.

To obtain that issue, the subtitles cues pushed should have overlapping
time boundaries in a specific way:
```
previous segment: |AAAAA|
new segment:         |BBBBBB|
```
Here the wanted result should have been:
```
Result:           |AAABBBBB|
```

But effectively became:
```
Result:              |BBBBB|
```

This was due to a mistake related to finding the index that new segment
should be pushed to in the corresponding segment array.

---

Real overlapping segments are very rare but this situation here happened
due to a very low rounding error when calculating a segment's end.

In the encountered case, the ending time of a segment N was offseted by a
fraction of a millisecond and was for this reason superior to the
starting time of the consecutive segment N+1.
Rounding errors happen and are generally handled properly so it's not a
real problem, it was just revealing here the real issue that this commit
fixes.