stream: fix logic for range eviction on QuotaExceededError
This is a very minor fix on what appeared to be an inefective strategy
when receiving a QuotaExceededError, which is an error which may appear
when a pushed segment overloads the memory.

The previous strategy was to try to clean-up ranges of data - that are
far away from the current position - from the buffer (a strategy
generally already performed by the browser).

However, and at least to me, the condition appeared to be the reverse of
what we should check for. We were not clearing far away ranges, but
ranges that were close (yet thankfully not the one containing the
current position).
Still, I welcome any reviewer to re-check, as this type of range-related
logic is notoriously hard to reason about.

We did not observe any bug or weird behavior linked to this problem and
this logic should very rarely be useful, so this is not a huge fix.