directfile: use getStartDate in getWallClockTime when available
This commit relies on `getStartDate`, when available, (for now just seen
on Safari when playing HLS contents) to provide a real "wall-clock time"
(which is in our case, the offseted position to correspond to the air
time) when playing directfile contents.

Previously, all directfile contents just returned the real non-offseted
position when calling `getWallClockTime`.