add exactOptionalPropertyTypes typescript option
This commit adds the `exactOptionalPropertyTypes` typescript option
which adds a semantic difference in property types and arguments
between the question mark (for optional parameters) and an union with
`undefined` (`| undefined`).

exactOptionalPropertyTypes documentation: https://www.typescriptlang.org/tsconfig#exactOptionalPropertyTypes

---

Both are generally equivalent without the option but with it, the
question mark only means that the property / argument may or may not be
present and `| undefined` will mean that the property / argument is
always present but may be set to `undefined`.

In some cases, such as when using the `for ... in` operator, checking a
function's arity or using methods like `hasOwnProperty`, there is a
difference of behavior between those two possibilities.

Moreover, the RxPlayer being a library used by several applications we do not
control, I prefer adding that supplementary distinction to limit even
more the possibility of type issues or regressions on them when they update the
RxPlayer's version.

---

For now, I mainly added in most places a `| undefined` when optional
property/arguments were defined, with a question mark.
Of course, this is not optimal as it adds a lot of clutter in the code.

We may want to write stricter code in the future to clearly chose one
between the two, and even update the current code so only one of both is
actually possible.
In types exposed to the API, I prefer however keeping the two
possibilities in most places for now as it simplifies the RxPlayer
maintainance.