drm: Use WebkitMediaKeys implementation on Safari Mobile
After a lot of tests today, we noticed that Safari mobile's standard
EME implementation worked in mysterious ways. Despite multiple retries,
we never could pass the `generateRequest` step, presumably due to a
non-satisfactory `initDataType` argument.

We ended up trying to use the webkit-prefixed APIs in its place, which
work differently than the standard but whose behavior is known.

And it worked! We could at last play encrypted directfile contents
(including HLS) on Safari IOS).

Note that we already did that for Safari Desktop due to similar issues
months (years?) ago, but we might not have tested properly on Safari iOS
(in directfile mode) at the time.

Because the `isSafari` constant was only `true` for Safari desktop, and
because I though making a difference between it and mobile could make
sense, I also renamed it into `isSafariDesktop`.