Fix typo which could lead to unintented gap skips in get_needed_segment.ts
While working on the future v4.0.0, I noticed an issue where
discontinuities created by either browser garbage collection or due to
the `SourceBuffer.remove` API would not be re-filled when re-encountered
but directly skipped instead.

This is because the current logic had a typo which led the RxPlayer to
ignore some segments GCed at their end.
Thankfully, instead of staying in rebuffering when that hole is
encountered, another module of the RxPlayer, the `StallAvoider`
understood that this hole won't be filled by any segment and directly
seeked over it when encountered