Add IBitrateEstimate, IPositionUpdate and IPlayerState types to the API
This commis adds to the exposed `rx-player/types` path the following
types:
  - IBitrateEstimate (payload of a `bitrateEstimationChange` event)
  - IPositionUpdate (payload of the `positionUpdate` event)
  - IPlayerState (return value of the `getPlayerState` method and
    payload of the `playerStateChange` event).

Those were missing and an issue was open (#1076) to ask us to add them.

I also profited from this commit to move type definition is
`src/core/api/public_api.ts` at the end of the file as we do now to
improve code readability when first reading it).

There's however a remaining problem in that VSCode (and I suppose
every other editors relying on tsserver's auto-import feature) seems to
prefer the real source file of the type definition on some auto-imports
and not the re-export done in the `rx-player/types` path (which is a
file generated at build time which re-exports everything from
`src/public_types.ts` which itself re-exports all API-exposed types in
one place).
This last issue leads to some types being auto-imported from deep inside
the RxPlayer code, which is moreover not part of the API.

Because it is a harder and completely different problem to solve
however, this is not fixed here.