Add "reload" audioTrackSwitchingMode
I was considering adding the "reload" audioTrackSwitchingMode as we
noticed that a specific Canal+ application (one linked to Media Asset
Management) experienced issues with the `"direct"`
`audioTrackSwitchingMode`, a mode allowing to switch between audio tracks
directly at the cost of a brief rebuffering period while the new audio track
is loaded - instead of the smoother default `"seamless"` we use for example
on the canalplus.com website, which still plays the old audio track
for a few seconds before switching.

Though I could not yet reproduce this issue on my side (due to security
mechanisms preventing me from testing it for now - though we're working
on it), I'm wondering if adding a `"reload"` mode in any case - which
would completely reload the `MediaSource` - would not make sense anyway.

The reason behind having this new mode would be to always provide the
`"reload"` solution at worst, which is usually the more compatible way to
change audio tracks.

Even if the `"direct"` mode was documented as possibly triggering a
RELOADING step, it was not actually possible yet (this was only added
to allow us to add it in the future without breaking the API). Thus it
can be splitted in two without breaking anything:
  - `"direct"`: only try to "flush" the buffer, usually through a small
    seek

  - `"reload"`: reload the MediaSource