Remove pinkie dependency
[Pinkie] was a Promise implementation used on platforms that didn't
support JavaScript's Promises, most notably, Internet Explorer 11.

It was used as a [ponyfill - and not polyfill -](https://ponyfoo.com/articles/polyfills-or-ponyfills)
so we wouldn't interfere with possible application-side polyfills.

However, there's multiple factors we didn't plan that made it a necessity
to still have a Promise polyfill in place (in which case pinkie is not
needed at all):
  1. Since quite some time, we added async/await syntax in some parts of
     the code. This transpiles down (depending on the target's ECMAScript
     version - but it's the case on our default builds) to JS code relying
     on native Promises by default.
  2. RxJS, which we used since forever, rely on native Promises for some
     logic we use

Applications generally had no problem relying on the RxPlayer on IE11
though because they already integrate, in the great majority, a Promise
polyfill when IE11-support is wanted (at least all the applications relying
on the RxPlayer on IE11 we know of did).

We actually knew that issue for some years, as it was already
reproducible on our standalone page on IE11, but did nothing yet due to
the complexity and low priority.

As we didn't get a single issue on it since then, it may be time to
officially remove the pinkie dependency.